chore: migrate to TypeScript
