+++ package.json
@@ -15,10 +15,15 @@
   "devDependencies": {
     "@cap-js/cds-test": "^0.4.0",
     "@cap-js/cds-types": "^0.13.0",
-    "@cap-js/sqlite": "^2"
+    "@cap-js/sqlite": "^2",
+    "@types/express": "^4",
+    "@types/node": "^22",
+    "ts-node": "^10.9.2",
+    "typescript": "^5"
   },
   "scripts": {
-    "start": "cds-serve",
-    "test": "node --test"
+    "build": "tsc",
+    "start": "npm run build && cds-serve",
+    "test": "npm run build && node --test test/*.test.js"
   }
 }
