--- test/basic.test.js
@@ -1,6 +1,0 @@
-const test = require('node:test');
-const assert = require('node:assert/strict');
-
-test('basic arithmetic works', () => {
-  assert.equal(2 + 2, 4);
-});
