+++ test/basic.test.ts
@@ -1,0 +1,6 @@
+import test from 'node:test';
+import assert from 'node:assert/strict';
+
+test('basic arithmetic works', () => {
+  assert.equal(2 + 2, 4);
+});
